feat: 新增语音测评功能

+++ src/ysw_web/routers/asr.py
@@ -1,41 +1,38 @@
-import uuid
-from typing import BinaryIO, List, Annotated, Literal
-from fastapi import APIRouter, File, UploadFile, Form, Request
-from starlette.responses import HTMLResponse
-from fastapi import File
-from fastapi.responses import JSONResponse
-import shutil
+from typing import Optional, List
 import os
-from faster_whisper import WhisperModel
-from loguru import logger
-from ysw_web.utils.response import success
-
+from fastapi import APIRouter, UploadFile, Form, Request
+from fastapi import File
 from ysw_ai.asr.factory import getASRModel
-
+from ysw_web.utils.download_utils import download_file, save_file
+from ysw_web.utils.response import success
 
 router = APIRouter()
 
 asr_model = getASRModel()
 
-TimestampGranularities = list[Literal["segment", "word"]]
-
 @router.post("/audio/transcriptions")
 async def transcriptions(
         request: Request,
-        file: UploadFile = File(...),
-        timestamp_granularities: Annotated[
-            TimestampGranularities,
-            Form(alias="timestamp_granularities[]"),
-        ] = ["segment"],
+        file: Optional[UploadFile] = File(None),
+        file_url: Optional[str] = Form(None),
+        language: Optional[str] = Form('en'),
+        timestamp_granularities: List[str] = Form(default=["segment"])
 ):
-    file_id = uuid.uuid4()
-    temp_file = f"temp_{file_id}_{file.filename}"
+
+    if not file and not file_url:
+        raise HTTPException(
+            status_code=400,
+            detail="必须提供文件上传或URL参数"
+        )
+    temp_file = None
     try:
-        # 保存上传的文件到临时文件
-        with open(temp_file, 'wb') as buffer:
-            shutil.copyfileobj(file.file, buffer)
+        if file:
+            temp_file = await save_file(file)
+        elif file_url:
+            temp_file = await download_file(file_url)
         # 使用 asr 模型进行转录
-        info = await asr_model.transcribe(temp_file, word_timestamps = "word" in timestamp_granularities)
+        info = await asr_model.transcribe(temp_file, language=language, word_timestamps = "word" in timestamp_granularities)
         return success(info)
     finally:
-        os.remove(temp_file)
+        if temp_file:
+            os.remove(temp_file)
